backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
httpx==0.28.1
python-multipart==0.0.22
jinja2==3.1.6
python-dotenv==1.2.2
anthropic
fpdf2

